updated to mask data on the fly. mostly operational now, too!

+++ bin/aei-palsar.py
@@ -16,39 +16,48 @@ import numpy as np
 import argparse
 
 # function to parse the input arguments
-class parse_args:
-    def __init__(self):
+def parse_args():
+    
+    # create the parser object
+    parser = argparse.ArgumentParser(description = "Processes ALOS-PALSAR data")
+    
+    # create options for input files
+    parser.add_argument('--hh', metavar = 'hh', type = str,
+        nargs = 1, help = 'path to the HH polarization data',
+        required = True)
+    
+    parser.add_argument('--hv', metavar = 'hv', type = str,
+        nargs = 1, help = 'path to the HV polarization data',
+        required = True)
         
-        # create the parser object
-        parser = argparse.ArgumentParser(description = "Processes ALOS-PALSAR data")
+    # create option for output file
+    parser.add_argument('-o', metavar = 'output_file', type = str,
+        nargs = 1, help = 'path to the output file',
+        required = True)
         
-        # create options for input files
-        parser.add_argument('--hh', metavar = 'hh', type = str,
-            nargs = 1, help = 'path to the HH polarization data',
-            required = True)
+    # create options for the despeckling algorithm to use
+    parser.add_argument('--filter', metavar = 'filter', type = str,
+        nargs = 1, help = 'the filter type',
+        choices = ['lee', 'frost', 'gammamap', 'kuan'], 
+        default = 'lee', required = False)
+    
+    # options for the despeckling radius    
+    parser.add_argument('--rad', metavar = 'radius', type = int,
+        nargs = 1, help = 'the despeckling radius', 
+        default = 3, required = False)
+    
+    # option to set the output nodata value    
+    parser.add_argument('--nodata', metavar = 'no data val', nargs = 1,
+        help = 'the output nodata value', default = -9999,
+        required = False)
         
-        parser.add_argument('--hv', metavar = 'hv', type = str,
-            nargs = 1, help = 'path to the HV polarization data',
-            required = True)
-            
-        # create option for output file
-        parser.add_argument('-o', metavar = 'output_file', type = str,
-            nargs = 1, help = 'path to the output file',
-            required = True)
-            
-        # create options for the despeckling algorithm to use
-        parser.add_argument('--filter', metavar = 'filter', type = str,
-            nargs = 1, help = 'the filter type',
-            options = ['lee', 'frost', 'gammamap', 'kuan'], 
-            default = 'lee', required = False)
-            
-        parser.add_argument('--rad', metavar = 'radius', type = int,
-            nargs = 1, help = 'the despeckling radius', 
-            default = 3, required = False)
-        
-        # parse the arguments and return the object
-        args = parser.parse_args()
-        return args
+    # option to not mask the output file
+    parser.add_argument('--nomask', action = 'store_true',
+        help = 'set to prevent masking', required = False)
+    
+    # parse the arguments and return the object
+    args = parser.parse_args()
+    return args
         
 # function to calculate the raw data to decibel
 def raw_to_decibel(infile, outfile):
@@ -57,7 +66,7 @@ def raw_to_decibel(infile, outfile):
     outfile = '"{}?&gdal:co:COMPRESS=LZW"'.format(outfile)
     
     # run the command
-    aei.cmd.otb.BandMath(infile, outfile, "10*log10(im1b1^2)-83")
+    aei.cmd.otb.BandMath(infile, outfile, '"10*log10(im1b1^2)-83"')
     
 # function to despeckle the data
 def despeckle(infile, outfile, filtername, radius):
@@ -78,7 +87,7 @@ def calc_polarimetry(hh, hv, outfile):
     outfile = '"{}?&gdal:co:COMPRESS=LZW"'.format(outfile)
     
     # build the expression string
-    exp = "(-1)/(im1b1+im2b1)*(im1b1*log2(im1b1/(im1b1+im2b1))+im2b1*log2(im2b1/(im1b1+im2b1)))"
+    exp = '"(-1)/(im1b1+im2b1)*(im1b1*log2(im1b1/(im1b1+im2b1))+im2b1*log2(im2b1/(im1b1+im2b1)))"'
     
     # run the command
     aei.cmd.otb.BandMath([hh, hv], outfile, exp)
@@ -91,6 +100,30 @@ def stack_outputs(hh, hv, pol, outfile):
     
     # run the command
     aei.cmd.otb.ConcatenateImages([hh, hv, pol], outfile)
+
+# function to build a mask from the input data 
+def get_mask(infile, outfile):
+    
+    # add compression to output file
+    outfile = '"{}?&gdal:co:COMPRESS=LZW"'.format(outfile)
+    
+    # run the command
+    cmd = "otbcli_ManageNoData -in {} -out {} uint8".format(
+        infile, outfile)
+    print(cmd)
+    aei.cmd.run(cmd)
+    
+# function to apply the mask
+def apply_mask(infile, outfile, maskfile, ndval):
+    
+    # add compression to output file
+    outfile = '"{}?&gdal:co:COMPRESS=LZW"'.format(outfile)
+    
+    # run the command
+    cmd = "otbcli_ManageNoData -in {} -out {} -mode apply -mode.apply.mask {} -mode.apply.ndval {}".format(
+        infile, outfile, maskfile, ndval)
+    print(cmd)
+    aei.cmd.run(cmd)
     
 # the main program function
 def main():
@@ -100,26 +133,61 @@ def main():
     
     # create three temp files for hh, hv, polarimetry data
     hh_dec = 'temp_HH_decibel.tif'
+    hh_msk = 'temp_HH_decibel_masked.tif'
     hh_des = 'temp_HH_despeckle.tif'
     hv_dec = 'temp_HV_decibel.tif'
-    hv_des = 'temp_HV_despeclkle.tif'
+    hv_msk = 'temp_HV_decibel_masked.tif'
+    hv_des = 'temp_HV_despeckle.tif'
     polari = 'temp_polarimetry.tif'
+    tmpmsk = 'temp_mask.tif'
+    premsk = 'temp_stack.tif'
     
-    # calculate the decibel and polarimetery data
-    raw_to_decibel(args.hh, hh_dec)
-    raw_to_decibel(args.hv, hv_dec)
+    # build a mask file, if set
+    if not args.nomask:
+        get_mask(args.hh[0], tmpmsk)
+    
+    # calculate the decibel data
+    raw_to_decibel(args.hh[0], hh_dec)
+    raw_to_decibel(args.hv[0], hv_dec)
+    
+    # mask the decibel data prior to despeckling
+    if not args.nomask:
+        apply_mask(hh_dec, hh_msk, tmpmsk, args.nodata)
+        apply_mask(hv_dec, hv_msk, tmpmsk, args.nodata)
+        os.remove(hh_dec)
+        os.remove(hv_dec)
+        hh_dec = hh_msk
+        hv_dec = hv_msk
+    
+    # despeckle the decible data
     despeckle(hh_dec, hh_des, args.filter, args.rad)
+    despeckle(hv_dec, hv_des, args.filter, args.rad)
+    
+    # calculate the polarimetry
     calc_polarimetry(hh_des, hv_des, polari)
     
-    # concatenate all files into a single stack
-    stack_outputs(hh_des, hv_des, polari, args.o)
+    # stack the outputs in different order if masking set
+    if args.nomask:
+        
+        # concatenate all files into a single stack
+        stack_outputs(hh_des, hv_des, polari, args.o[0])
+        
+    else:
+        
+        # concatenate all files into a single stack
+        stack_outputs(hh_des, hv_des, polari, premsk)
+        
+        # then apply the mask
+        apply_mask(premsk, args.o[0], tmpmsk, args.nodata)
     
     # delete the temp files
-    os.remove(hh_dec)
-    os.remove(hv_dec)
-    os.remove(hh_des)
-    os.remove(hv_des)
-    os.remove(polari)
+    #os.remove(hh_dec)
+    #os.remove(hv_dec)
+    #os.remove(hh_des)
+    #os.remove(hv_des)
+    #os.remove(polari)
+    #os.remove(tmpmsk)
+    #os.remove(premsk)
     
     # report finished
     print("[ STATUS ]: Finished ALOS-PALSAR processing!")
